Sorry for monolithic commit. Will break it up in the future.
Moved the external libraries to separate library files so that libcantera.a just contains its own namespace externals.

Fixed several errors in the equilibrium program that occurred during the port. (int to size_t issues).

Moved some equilibrium program headers to the include file system, so that it can link with equilibrium program.

Worked on Cantera.mak. Needs more work.

Fixed an issue with the Residual virtual base classes within numerics. They didn't inherit due to int to size_t migration. This caused numerous test problems to fail (issue with backwards compatibility - do we want it and how much do we want?).

Added csvdiff back so that it's available for shell environment runtests.

+++ cantera/trunk/src/numerics/ResidJacEval.cpp
@@ -312,7 +312,7 @@ matrixConditioning(doublereal* const matrix, const int nrows, doublereal* const 
 int ResidJacEval::
 evalResidNJ(const doublereal t, const doublereal deltaT,  const doublereal* y,
             const doublereal* ydot, doublereal* const resid, const ResidEval_Type_Enum evalType,
-            const size_t id_x, const doublereal delta_x)
+            const int id_x, const doublereal delta_x)
 {
     throw CanteraError("ResidJacEval::evalResidNJ()", "Not implemented\n");
     return 1;
